add CKFinderAdapter file
